Home work 5

+++ controllers/EventController.php
@@ -3,11 +3,17 @@
 namespace app\controllers;
 
 use Yii;
+use app\models\forms\EventForm;
 use app\models\Event;
+use app\models\Access;
 use app\models\EventSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use app\objects\ViewModels\EventCreateView;
+use app\objects\ViewModels\EventView;
+use yii\web\ForbiddenHttpException;
 
 /**
  * EventController implements the CRUD actions for Event model.
@@ -29,11 +35,17 @@ class EventController extends Controller
                     'delete' => ['POST'],
                 ],
             ],
-            'access' => AccessControl::class,
-            'only'
-            'rules' => [
-                
-            ]
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['create', 'index', 'update', 'delete'],
+                'rules' => [
+                    [
+                        'allow' =>true,
+                        'roles' =>['@'],
+                    ],
+                ],
+            
+            ],
         ];
     }
 
@@ -60,8 +72,15 @@ class EventController extends Controller
      */
     public function actionView($id)
     {
+        $event = $this->findModel($id);
+        $viewModel = new EventView();
+
+        if (!$this->checkAccess($event)) {
+            throw new ForbiddenHttpException('У вас нет доступа к текущему событию.');
+        }
         return $this->render('view', [
-            'model' => $this->findModel($id),
+            'model' => $event,
+            'viewModel' => $viewModel,
         ]);
     }
 
@@ -72,19 +91,24 @@ class EventController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Event();
-
-        if (Yii::$app->request->post()) {
-            $incomingData = Yii::$app->request->post();
-            $incomingData['Event']['uid'] = (Yii::$app->user)->id;
-
-            if ($model->load($incomingData) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
-            }
+        $model = new EventForm();
+        /* $model->load(Yii::$app->request->post()); */
+        /* var_dump($model->save()); */
+        /* die; */
+        /* $model->uid =10; */
+        /* if (Yii::$app->request->post()) { */
+        /*     var_dump(Yii::$app->request->post()); */
+        /*     die; */
+        
+        /* } */
+        if ($model->load(Yii::$app->request->post()) && $model->save() && $model->validate()) {
+            return $this->redirect(['view', 'id' => $model->id]);
         }
 
+        $viewModel = new EventCreateView();
         return $this->render('create', [
             'model' => $model,
+            'viewModel' => $viewModel,
         ]);
     }
 
@@ -99,12 +123,24 @@ class EventController extends Controller
     {
         $model = $this->findModel($id);
 
+
+        if (!$this->checkWriteAccess($model)) {
+            throw new ForbiddenHttpException('У вас нет прав редактировать это событие.');
+        }
+
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
+
+        if (!$this->checkTimeAccess($model)) {
+        
+            throw new ForbiddenHttpException('Нельзя редактировать завершенные события.');
+        };
             return $this->redirect(['view', 'id' => $model->id]);
         }
 
+        $viewModel = new EventCreateView();
         return $this->render('update', [
             'model' => $model,
+            'viewModel' => $viewModel,
         ]);
     }
 
@@ -117,6 +153,11 @@ class EventController extends Controller
      */
     public function actionDelete($id)
     {
+        $event = $this->findModel($id);
+        if (!$this->checkWriteAccess($event)) {
+            throw new ForbiddenHttpException('У вас нет прав редактировать это событие.');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -131,10 +172,39 @@ class EventController extends Controller
      */
     protected function findModel($id)
     {
-        if (($model = Event::findOne($id)) !== null) {
+        if (($model = EventForm::findOne($id)) !== null) {
             return $model;
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
+    }
+
+    protected function checkAccess(Event $event): bool
+    {
+        $curentUid = \Yii::$app->user->getId();
+        if ($event->uid == $curentUid) {
+            return true;
+        } elseif (Access::find()->andWhere([
+            'eventid' => $event->id,
+            'userid' => $curentUid
+        ])->count()) {
+            return true;
+        }
+        return false;
+    }
+
+    protected function checkWriteAccess(Event $event)
+    {
+        return $event->uid == \Yii::$app->getUser()->getId();
+    }
+
+    protected function checkTimeAccess(EventForm $event) 
+    {
+        $timeevent = strtotime($event->end_at);
+        $today = strtotime(date("Y-m-d H:i:s"));
+        $out = $timeevent > $today;
+
+        return $out;
+    
     }
 }
